move calculate_steps_per_epoch
